Fixed property types.

+++ TSMDotNetApi/Models/TsmItemGlobalData.cs
@@ -7,14 +7,14 @@
         public int EuQuantity { get; set; }
         public int EuHistoricalPrice { get; set; }
         public int EuSaleAvg { get; set; }
-        public int EuAvgDailySold { get; set; }
-        public int EuSaleRate { get; set; }
+        public float EuAvgDailySold { get; set; }
+        public float EuSaleRate { get; set; }
         public int UsMarketAvg { get; set; }
         public int UsMinBuyoutAvg { get; set; }
         public int UsQuantity { get; set; }
         public int UsHistoricalPrice { get; set; }
         public int UsSaleAvg { get; set; }
-        public int UsAvgDailySold { get; set; }
-        public int UsSaleRate { get; set; }
+        public float UsAvgDailySold { get; set; }
+        public float UsSaleRate { get; set; }
     }
 }
